Update Slice.js
It looks like there's a small typo in your Redux reducer. The issue is with the updateClickValue reducer in your notesSlice. You are adding a new note with Text property, but later when you are mapping over the value in your Notes component, you are using each.text to access the text property. The correct property name is Text.